+++ src/components/FormValidator.js
@@ -34,18 +34,6 @@ export class FormValidator {
         const isValid = inputList.some((inputElement, i) => {
             return !inputElement.validity.valid;
         });
-        if (inputList[2]===undefined) return isValid
-        if (isValid) {
-            return true
-        } else if (!isValid && ((inputList[1].value === inputList[2].value))) {
-            
-            return false
-        } else {
-            const inputElement = document.querySelector("#third-field-place")
-            const errorMessage = 'Passwords are not the same'
-            this.showInputError(inputElement, errorMessage)
-            return true
-        }
     }
     
     _toggleButtonState(inputList) {
