+++ src/index.html
@@ -15,7 +15,7 @@
         content="Everything you need to practice SSAT and ISEE effectively. Based on Greenberg School materials.">
   <meta name="keywords" content="SSAT, ISEE, test, preparation, prep, math, analogy">
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-<!--  <link rel="stylesheet" type="text/css" href="pages/index.css">-->
+  <link rel="stylesheet" type="text/css" href="pages/index.css">
   <title>SSAT and ISEE Test Preparation | tutor4exam</title>
 </head>
 <body>
@@ -276,6 +276,6 @@
 <script src="https://www.gstatic.com/firebasejs/7.23.0/firebase-app.js"></script>
 <script src="https://www.gstatic.com/firebasejs/7.23.0/firebase-auth.js"></script>
 <script src="https://www.gstatic.com/firebasejs/7.23.0/firebase-firestore.js"></script>
-<!--<script type="module" src="pages/index.js"></script>-->
+<script type="module" src="pages/index.js"></script>
 </body>
 </html>
