+++ src/utils/glancecards.js
@@ -1,22 +1,25 @@
-import firstImg from './../images/choose1.webp';
-import secondImg from './../images/choose2.webp';
-import thirdImg from './../images/choose3.webp';
+// import firstImg from './../images/choose1.webp';
+// import secondImg from './../images/choose2.webp';
+// import thirdImg from './../images/choose3.webp';
 
 export const glanceCards = [
     {
-        link: firstImg,
+        // link: firstImg,
+        link: './images/choose1.webp',
         alt: "Square anthropomorphic characters playing a math game",
         title: "Math and Analogies",
         text: "More than 500 SSAT & ISEE test questions and problems with tutorials based on 10 levels of difficulty"
     },
     {
-        link: secondImg,
+        // link: secondImg,
+        link: './images/choose1.webp',
         alt: "Square character shredding list to digits",
         title: "Different topics",
         text: "More than 100 topics for students to learn and excel at."
     },
     {
-        link: thirdImg,
+        // link: thirdImg,
+        link: './images/choose1.webp',
         alt: "Square character trying to solve problem",
         title: "Help is here",
         text: "More than 100 clues to guide you to a better way to solve a problem"
